fix migration error

+++ database/migrations/1625722076444_m_kategori_pelanggaran_schema.js
@@ -5,17 +5,20 @@ const Schema = use("Schema");
 
 class MKategoriPelanggaranSchema extends Schema {
   up() {
-    this.create("m_kategori_pelanggaran", (table) => {
-      table.increments();
-      table.string("nama");
-      table.integer("m_sekolah_id").unsigned().index("m_sekolah_id");
-      table
-        .foreign("m_sekolah_id")
-        .references("m_sekolah.id")
-        .onDelete("cascade")
-        .onUpdate("cascade");
-      table.boolean("dihapus").defaultTo(0);
-      table.timestamps();
+    this.hasTable("m_kategori_pelanggaran").then((exist) => {
+      if (!exist)
+        this.create("m_kategori_pelanggaran", (table) => {
+          table.increments();
+          table.string("nama");
+          table.integer("m_sekolah_id").unsigned().index("m_sekolah_id");
+          table
+            .foreign("m_sekolah_id")
+            .references("m_sekolah.id")
+            .onDelete("cascade")
+            .onUpdate("cascade");
+          table.boolean("dihapus").defaultTo(0);
+          table.timestamps();
+        });
     });
   }
